requirements (1).txt

streamlit
sounddevice
wavio
SpeechRecognition
gTTS
python-dotenv
google-generativeai
